agent/requirements.txt
gspread==5.12.4
google-auth==2.28.1
google-search-results==2.4.2
google-generativeai==0.7.2
langchain-google-genai>=1.0.1
langgraph>=0.0.38
python-dotenv>=1.0.1
pytest>=8.1.1
pytest-mock>=3.12.0
pandas>=2.2.1
pandas-ta>=0.3.14b0
tenacity>=8.2.3
polygon-api-client>=1.12.0
sentry-sdk>=1.45.0
rich>=13.7.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
langchain>=0.1.5
langchain-core>=0.1.15
fredapi>=0.5.1
scipy>=1.11.0
requests>=2.31.0
